feat(create-plugin): add panel-docs codemod with AI authoring assistance
