+++ packages/create-plugin/templates/docs/common/workflows/validate-docs.yml
