+++ packages/create-plugin/templates/docs/panel/docs/examples.md
