+++ packages/create-plugin/templates/docs/panel/docs/index.md
