+++ packages/create-plugin/templates/docs/panel/docs/options.md
