+++ packages/create-plugin/templates/docs/panel/docs/troubleshooting.md
